pyproject.toml
[project]
name = "news"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.6"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "lxml>=6.0.2",
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
